fix(test): update  model class name after rename
