+++ backend/sgset/tests/unit/models/test_course.py
@@ -1,14 +1,14 @@
 import pytest
 from decimal import Decimal
-from sgset.models import CourseModel
+from sgset.models import Course
 import uuid
 
 
 @pytest.mark.django_db
 @pytest.fixture
 def course_instance():
-    """Cria uma instância de CourseModel para os testes."""
-    return CourseModel.objects.create(
+    """Cria uma instância de Course para os testes."""
+    return Course.objects.create(
         name="Mechanical Fitter",
         price_per_student=Decimal("1200.00")
     )
@@ -16,7 +16,7 @@ def course_instance():
 
 @pytest.mark.django_db
 def test_course_creation(course_instance):
-    """Teste a criação de uma instância de CourseModel."""
+    """Teste a criação de uma instância de Course."""
     course = course_instance
     assert isinstance(course.id, uuid.UUID)  # Verifica se o id é um UUID válido
     assert course.name == "Mechanical Fitter"
@@ -25,14 +25,14 @@ def test_course_creation(course_instance):
 
 @pytest.mark.django_db
 def test_course_str(course_instance):
-    """Teste a string de representação do CourseModel."""
+    """Teste a string de representação do Course."""
     course = course_instance
     assert str(course) == "Mechanical Fitter - $1200.00"
     
 @pytest.mark.django_db
 def test_course_without_price():
-    """Teste a criação de CourseModel sem price_per_student."""
-    course = CourseModel.objects.create(
+    """Teste a criação de Course sem price_per_student."""
+    course = Course.objects.create(
         name="Arduino"
     )
     assert course.name == "Arduino"
