+++ backend/sgset/tests/unit/models/test_day_of_week.py
@@ -1,12 +1,12 @@
 import pytest
-from sgset.models import DayOfWeekModel
+from sgset.models import DayOfWeek
 import uuid
 
 
 @pytest.mark.django_db
 @pytest.fixture
 def day_of_week_instance():
-    return DayOfWeekModel.objects.create(name="Monday")
+    return DayOfWeek.objects.create(name="Monday")
 
 
 @pytest.mark.django_db
@@ -24,6 +24,6 @@ def test_day_of_week_str(day_of_week_instance):
 
 @pytest.mark.django_db
 def test_unique_day_of_week():
-    DayOfWeekModel.objects.create(name="Tuesday")
+    DayOfWeek.objects.create(name="Tuesday")
     with pytest.raises(Exception):
-        DayOfWeekModel.objects.create(name="Tuesday")
+        DayOfWeek.objects.create(name="Tuesday")
